requirements.txt
torch
torchvision
nltk
jieba
tqdm
transformers
peft
bert4keras
scikit-learn
loguru
python-dotenv
openai

